add `npm run dev`
- dev mode is started by calling webpack directly (no scripts/dev.js)
because when webpack.compiler.watch is called from within a script it
failed to watch for file changes.
- Webpack.config was updated so that it always outputs assets that are
needed for django (rather than have them handled by a loader in
webpackDevServer).
- a quick and dirty webpack plugin to generate django templates was
added. This means that hashes could be added back to the static asset
- a flag in settings.py controls whether the app is in production (i.e.
a user of the app) vs development (i.e. working on datavalidation app
itself), this is needed to that prod and dev static assets/templates
can be kept separate. This is a bit ugly, so thinking of a better
solution is a todo