don't display num_passing=0 when it is really undefined
If num passing was not defined (e.g. because the data_validator only
return a queryset with the failing objects) then we shouldn't show 0
in the num_passing field on the front end. In stead add some logic to
SummaryEx so that we can make num_passing, num_failing, and num_na None
if they weren't specified by the user