add DataValidationMixin
lets users run validators in the django admin change forms. For this
to work the FailingObject models needs the content_type field. Ideally
this could be specified only on the Validator model and the
GenericRelation on FailingObject could use the foreign key to validator
instead of having its own content_type field. That might require some
more work.